app/requirements.txt
absl-py
asttokens
astunparse
blinker
cachetools
certifi
charset-normalizer
click
colorama
comm
debugpy
decorator
exceptiongroup
executing
Flask
flatbuffers
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib-metadata
ipykernel
ipython
itsdangerous
jedi
Jinja2
joblib
jupyter_client
jupyter_core
keras
libclang
Markdown
MarkupSafe
matplotlib-inline
ml-dtypes
mysql-connector
mysql-connector-python
nest-asyncio
numpy
oauthlib
opt-einsum
packaging
pandas
parso
platformdirs
prompt-toolkit
protobuf
psutil
pure-eval
pyasn1
pyasn1-modules
Pygments
python-dateutil
pytz
pyzmq
requests
requests-oauthlib
rsa
scikit-learn
scipy
six
stack-data
tensorboard
tensorboard-data-server
tensorflow
termcolor
threadpoolctl
tornado
traitlets
typing_extensions
tzdata
urllib3
wcwidth
Werkzeug
wrapt
zipp
